refactor: centralize Redis client protocol and improve auth test assertions
Address code review feedback by creating a shared Redis module and
enhancing test robustness.

### Redis Client Refactoring
- Create backend/app/core/redis.py with shared RedisClientProtocol
- Add get_redis_client() factory to centralize redis.from_url() calls
- Consolidate all type: ignore[no-untyped-call] annotations in one place
- Update alert_service.py, celery/database.py, main.py, and admin.py
  to use shared protocol and factory
- Remove duplicate RedisClientProtocol definitions

### Auth Test Improvements
- Add specific error message assertions to auth tests (13 tests updated)
- Tests now verify both status code AND detail message:
  - 401: "Not authenticated" (missing/invalid credentials)
  - 403: "Admin privileges required" (authenticated but unauthorized)
- Makes tests more robust against regressions in auth layer
- Updated tests in: test_admin_alerts.py, test_admin_dashboard.py,
  test_auth_integration.py, test_tfl_api.py

### Test Fixes
- Update test_main.py to mock app.main.get_redis_client instead of
  app.main.redis.from_url
- Update test_alert_service.py imports for new module structure

All 1415 tests passing, 95.61% coverage, mypy/ruff passing.

Resolves code review comments on PR #281.
