+++ backend/tests/test_admin_alerts.py
@@ -57,7 +57,7 @@ async def test_trigger_check_requires_auth(async_client_with_db: AsyncClient) ->
     """Test that trigger-check endpoint requires authentication."""
     response = await async_client_with_db.post(build_api_url("/admin/alerts/trigger-check"))
     assert response.status_code == 401  # FastAPI 0.122+ returns 401 for missing credentials per RFC 7235
-    assert "detail" in response.json()
+    assert response.json()["detail"] == "Not authenticated"
 
 
 @pytest.mark.asyncio
@@ -80,7 +80,7 @@ async def test_worker_status_requires_auth(async_client_with_db: AsyncClient) ->
     """Test that worker-status endpoint requires authentication."""
     response = await async_client_with_db.get(build_api_url("/admin/alerts/worker-status"))
     assert response.status_code == 401  # FastAPI 0.122+ returns 401 for missing credentials per RFC 7235
-    assert "detail" in response.json()
+    assert response.json()["detail"] == "Not authenticated"
 
 
 @pytest.mark.asyncio
@@ -103,7 +103,7 @@ async def test_recent_logs_requires_auth(async_client_with_db: AsyncClient) -> N
     """Test that recent-logs endpoint requires authentication."""
     response = await async_client_with_db.get(build_api_url("/admin/alerts/recent-logs"))
     assert response.status_code == 401  # FastAPI 0.122+ returns 401 for missing credentials per RFC 7235
-    assert "detail" in response.json()
+    assert response.json()["detail"] == "Not authenticated"
 
 
 @pytest.mark.asyncio
