+++ backend/tests/test_auth_integration.py
@@ -85,12 +85,12 @@ class TestAuthMeEndpoint:
     """Tests for GET /auth/me endpoint."""
 
     def test_get_me_without_auth(self) -> None:
-        """Test /auth/me returns 403 without authentication."""
+        """Test /auth/me returns 401 without authentication."""
         with TestClient(app) as client:
             response = client.get("/api/v1/auth/me")
 
             assert response.status_code == 401  # FastAPI 0.122+ returns 401 for missing credentials per RFC 7235
-            assert "detail" in response.json()
+            assert response.json()["detail"] == "Not authenticated"
 
     def test_get_me_with_invalid_token(self) -> None:
         """Test /auth/me returns 401 with invalid token."""
@@ -103,13 +103,13 @@ class TestAuthMeEndpoint:
             assert "detail" in response.json()
 
     def test_get_me_with_malformed_authorization_header(self) -> None:
-        """Test /auth/me returns 403 with malformed Authorization header."""
+        """Test /auth/me returns 401 with malformed Authorization header."""
         with TestClient(app) as client:
             # Missing 'Bearer' prefix
             headers = {"Authorization": "NotBearerToken"}
             response = client.get("/api/v1/auth/me", headers=headers)
             assert response.status_code == 401  # FastAPI 0.122+ returns 401 for missing credentials per RFC 7235
-            assert "detail" in response.json()
+            assert response.json()["detail"] == "Not authenticated"
 
     def test_get_me_with_expired_token(self) -> None:
         """Test /auth/me returns 401 with expired JWT token."""
